Boss!
BIG GOOB
- fixed difficulty multiplier (sometimes had weird values)

+++ Cannoon/Assets/Scripts/Enemy/FlyingEnemyAI.cs
@@ -61,7 +61,7 @@ public class FlyingEnemyAI : MonoBehaviour
 
     private void ApplyDifficultyRating()
     {
-        currentSpeed = baseSpeed * endlessModeScript.difficultyMultiplier;
+        currentSpeed = (float)(baseSpeed * endlessModeScript.difficultyMultiplier);
     }
 
     // Update is called once per frame
